refactor: add interface file suffix
