+++ plugin-server/src/worker/ingestion/groups/repositories/raw-postgres-group-repository.interface.ts
